requirements.txt
python-bitcoinrpc>=1.0.0 